requirements.txt
python-for-android==2023.9.4
Kivy==2.2.1
kivymd==1.1.1
buildozer==1.5.0 